+++ .eleventy.js
@@ -3,6 +3,16 @@ import { eleventyImageTransformPlugin } from "@11ty/eleventy-img";
 import pluginSlideDeck from "./index.js";
 import yaml from "js-yaml";
 
+import markdownIt from 'markdown-it';
+import mdMark from 'markdown-it-mark';
+
+const mdOverride = markdownIt({
+    html: true,
+    breaks: false,
+    typographer: true,
+  }).disable('code')
+  .use(mdMark);
+
 const buildFunction = (slide) => {
   if (slide.youtube) {
     const bg = `background-image: url('https://v1.opengraph.11ty.dev/https%3A%2F%2Fyoutube.com%2Fwatch%3Fv%3D${slide.youtube}/auto/jpeg/');`;
@@ -24,7 +34,14 @@ export default async function(eleventyConfig) {
     imgDir: '/_img/',
 
     buildFunction,
+    markdownFunctions: {
+      block: (content) => mdOverride.render(content),
+      inline: (content) => mdOverride.renderInline(content),
+    }
   });
+
+  eleventyConfig.setLibrary('md', mdOverride);
+  eleventyConfig.addDataExtension("yaml", (contents) => yaml.load(contents));
 
   eleventyConfig.addPlugin(pluginWebc, {
     components: [
@@ -32,8 +49,6 @@ export default async function(eleventyConfig) {
       'test/_includes/**/*.webc',
     ],
   });
-
-  eleventyConfig.addDataExtension("yaml", (contents) => yaml.load(contents));
 
   eleventyConfig.addPlugin(eleventyImageTransformPlugin, {
     // output image formats