+++ index.js
@@ -17,9 +17,14 @@ export default async function(eleventyConfig, options) {
       typographer: true,
     },
 
+    // markdownFunctions: {
+    //   inline: mdDefault.renderInline(),
+    //   block: mdDefault.render(),
+    // },
+
     // imgDir: (path relative to content folder),
     // buildFunction: (function),
-    // newPenTemplates: { new: 'https://pen.new' }
+
     collectionName: 'slideDeck',
     known: {
       slides: 'knownSlides',
@@ -34,6 +39,15 @@ export default async function(eleventyConfig, options) {
     },
   }, options);
 
+  if (!options.markdownFunctions) {
+    const mdDefault = markdownIt(options.markdownIt).disable('code');
+
+    options.markdownFunctions = {
+      inline: (content) => mdDefault.renderInline(content),
+      block: (content) => mdDefault.render(content),
+    };
+  }
+
   // data
   eleventyConfig.addGlobalData("slideDeckConfig", options);
 
@@ -46,15 +60,14 @@ export default async function(eleventyConfig, options) {
 
   // markdown
   eleventyConfig.addPlugin(syntaxHighlightPlugin);
-  const mdIt = markdownIt(options.markdownIt).disable('code');
 
   eleventyConfig.addFilter(
-    'slideMDownBlock',
-    (content) => mdIt.render(content)
+    'slideMarkdownBlock',
+    options.markdownFunctions.block
   );
   eleventyConfig.addFilter(
-    'slideMDownInline',
-    (content) => mdIt.renderInline(content)
+    'slideMarkdownInline',
+    options.markdownFunctions.inline
   );
 
   // collection